Complete implementation of service-search, the combination of all munch-core search

+++ catalyst-munch/src/main/java/munch/catalyst/data/Location.java
@@ -23,7 +23,7 @@ public final class Location {
     private String center;
     private List<String> points;
 
-    private int sort;
+    private long sort;
     private Date updatedDate;
 
     public String getId() {
@@ -74,11 +74,11 @@ public final class Location {
         this.center = center;
     }
 
-    public int getSort() {
+    public long getSort() {
         return sort;
     }
 
-    public void setSort(int sort) {
+    public void setSort(long sort) {
         this.sort = sort;
     }
